Update to Raspbian 2018-04-18

+++ src/build-rpi2-qemu.sh
@@ -4,14 +4,30 @@ set -exm
 
 # User configurable variables
 KERNEL_IMAGE="kernel-qemu-4.9.59-stretch"
-RASPBIAN_DATE="2018-03-13"
-RASPBIAN_URL="http://director.downloads.raspberrypi.org/raspbian/images/raspbian-2018-03-14"
-RASPBIAN_MIRROR_URL="http://ftp.jaist.ac.jp/pub/raspberrypi/raspbian/images/raspbian-2018-03-14"
+KERNEL_SHA1=bdba230db1d404b41485a002d60b0db4ba571f1d
+DTB=versatile-pb.dtb
+DTB_SHA1=ab13a35a2b0862ee769a9de20bc975104f4f244d
+RASPBIAN_DATE="2018-04-18"
+RASPBIAN_URL="http://director.downloads.raspberrypi.org/raspbian/images/raspbian-2018-04-19"
+RASPBIAN_MIRROR_URL="http://ftp.jaist.ac.jp/pub/raspberrypi/raspbian/images/raspbian-2018-04-19"
+RASPBIAN_ZIP_SHA1=f141331a5613c3d5d94ff60cdbf0a25097b3520a
 
 SSH="ssh -p 2222 -o StrictHostKeyChecking=no"
 SCP="scp -P 2222 -o StrictHostKeyChecking=no"
 RASPBIAN_ZIP=${RASPBIAN_DATE}-raspbian-stretch.zip
 RASPBIAN_IMAGE=${RASPBIAN_DATE}-raspbian-stretch.img
+
+check_sha1()
+{
+    file=`mktemp`
+    cat >> $file <<EOF
+    ${KERNEL_SHA1} ${KERNEL_IMAGE}
+    ${DTB_SHA1} ${DTB}
+    ${RASPBIAN_ZIP_SHA1} ${RASPBIAN_ZIP}
+EOF
+    sha1sum -c $file
+    rm $file
+}
 
 die()
 {
@@ -45,27 +61,42 @@ ssh_control()
 EOF
     ${SCP} script pi@localhost:
     trap die ERR
-    ${SSH} pi@localhost bash script
+    ${SSH} -t pi@localhost bash script
     ${SCP} pi@localhost:artisan/src/\*.deb src
     pkill qemu-system-arm
 }
 
 ssh-keygen -R "[localhost]:2222"
-(sleep 600; pgrep curl && pkill curl)&
-curl -L -O ${RASPBIAN_URL}/${RASPBIAN_ZIP} || curl -L -O ${RASPBIAN_MIRROR_URL}/${RASPBIAN_ZIP}
-unzip ${RASPBIAN_ZIP}
-curl -L -O https://github.com/juokelis/qemu-rpi-kernel/raw/master/${KERNEL_IMAGE}
-curl -L -O https://github.com/juokelis/qemu-rpi-kernel/raw/master/versatile-pb.dtb
+if [ ! -f ${RASPBIAN_ZIP} ]; then
+    (sleep 600; pgrep curl && pkill curl)&
+    curl -L -O ${RASPBIAN_URL}/${RASPBIAN_ZIP} || curl -L -O ${RASPBIAN_MIRROR_URL}/${RASPBIAN_ZIP}
+
+fi
+if [ ! -f ${RASPBIAN_IMAGE} ]; then
+    unzip ${RASPBIAN_ZIP}
+fi
+if [ ! -f ${KERNEL_IMAGE} ]; then
+    curl -L -O https://github.com/juokelis/qemu-rpi-kernel/raw/master/${KERNEL_IMAGE}
+fi
+if [ ! -f versatile-pb.dtb ]; then
+    curl -L -O https://github.com/juokelis/qemu-rpi-kernel/raw/master/versatile-pb.dtb
+fi
+check_sha1
 qemu-img resize ${RASPBIAN_IMAGE} +2G
+LABEL_ID=`sfdisk -d ${RASPBIAN_IMAGE} | grep label-id | awk -F: '{ print $2 }'`
+if [ -z $LABEL_ID ]; then
+    echo unable to read label-id
+    exit 1
+fi
 partitions=`mktemp`
 cat <<EOF > $partitions
 label: dos
-label-id: 0x15ca46a5
-device: 2018-03-14-raspbian-stretch.img
+label-id: ${LABEL_ID}
+device: ${RASPBIAN_IMAGE}
 unit: sectors
 
-2018-03-14-raspbian-stretch.img1 : start=        8192, size=       85611, type=c
-2018-03-14-raspbian-stretch.img2 : start=       98304, size=    13860863, type=83
+${RASPBIAN_IMAGE}1 : start=        8192, size=       85611, type=c
+${RASPBIAN_IMAGE}2 : start=       98304, size=    13860863, type=83
 EOF
 sfdisk  ${RASPBIAN_IMAGE} < $partitions
 rm $partitions
@@ -79,6 +110,8 @@ sudo sh -c "echo 'systemctl set-default multi-user.target --force' >> $mountpoin
 sudo sh -c "echo 'systemctl stop lightdm.service --force' >> $mountpoint/etc/rc.local"
 sudo sh -c "echo 'systemctl stop graphical.target --force' >> $mountpoint/etc/rc.local"
 sudo sh -c "echo 'systemctl stop plymouth.service --force' >> $mountpoint/etc/rc.local"
+# HACK to work around piwheels.org DNS problems
+sudo sh -c "echo '93.93.129.174 proxy.mythic-beasts.com piwheels.org www.piwheels.org' >> $mountpoint/etc/hosts"
 sudo mkdir $mountpoint/home/pi/.ssh
 sudo chown 1000  $mountpoint/home/pi/.ssh
 sudo chmod go-rwx $mountpoint/home/pi/.ssh
